+++ backend/server/database/model/shared_network_test.go
@@ -221,7 +221,7 @@ func TestUpdateSharedNetwork(t *testing.T) {
 	returned, err = GetSharedNetwork(db, network.ID)
 	require.NoError(t, err)
 	require.NotNil(t, returned)
-	require.EqualValues(t, 10, returned.Utilization)
+	require.EqualValues(t, 10, returned.AdrUtilization)
 	require.EqualValues(t, 20, returned.PdsUtilization)
 }